requirements.txt
asgiref==3.9.1
bleach==6.2.0
dj-database-url==3.0.1
Django==5.2.4
django-axes==8.0.0
django-csp==4.0
gunicorn==23.0.0
Markdown==3.8.2
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.1
webencodings==0.5.1
whitenoise==6.9.0
coverage==7.10.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
drf-spectacular==0.28.0
djangorestframework-camel-case==1.4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-cors-headers==4.4.0
